fix: Moved client code to src folder + server serving correct static files.
