+++ client/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDom from 'react-dom'
-import App from './components/App'
-import { socketClientHandlers } from './utils/SocketClientHandlers'
+import App from './src/components/App'
+import { socketClientHandlers } from './src/utils/SocketClientHandlers'
 import io from 'socket.io-client'
 
 //Redux:
@@ -15,7 +15,7 @@ import {
 } from '../shared/src/constants/SOCKET_IO_DISPATCHERS'
 
 import { I18nextProvider } from 'react-i18next'
-import i18n from './i18n'
+import i18n from './src/utils/i18n'
 import { IMixerProtocol } from '../shared/src/constants/MixerProtocolInterface'
 
 declare global {
