+++ client/src/assets/css/ChanStrip.css
