+++ client/src/assets/css/ChanStripEq.css
