+++ client/src/assets/css/ChanStripFull.css
