+++ client/src/assets/css/MiniChanStrip.css
