+++ client/src/assets/css/ReductionMeter.css
