+++ client/src/components/ChanStripEq.tsx
@@ -5,10 +5,10 @@ import ReactSlider from 'react-slider'
 import '../assets/css/ChanStripEq.css'
 import { Store } from 'redux'
 import { connect } from 'react-redux'
-import { IFader } from '../../shared/src/reducers/fadersReducer'
-import { IChannel } from '../../shared/src/reducers/channelsReducer'
-import { SOCKET_SET_FX } from '../../shared/src/constants/SOCKET_IO_DISPATCHERS'
-import { fxParamsList } from '../../shared/src/constants/MixerProtocolInterface'
+import { IFader } from '../../../shared/src/reducers/fadersReducer'
+import { IChannel } from '../../../shared/src/reducers/channelsReducer'
+import { SOCKET_SET_FX } from '../../../shared/src/constants/SOCKET_IO_DISPATCHERS'
+import { fxParamsList } from '../../../shared/src/constants/MixerProtocolInterface'
 import { getFaderLabel } from '../utils/labels'
 
 interface IChanStripFullInjectProps {
