+++ client/src/components/ChanStripFull.tsx
@@ -8,18 +8,18 @@ import {
     storeShowOptions,
     storeShowMonitorOptions,
     storeShowChanStripFull,
-} from '../../shared/src/actions/settingsActions'
-import { IFader } from '../../shared/src/reducers/fadersReducer'
+} from '../../../shared/src/actions/settingsActions'
+import { IFader } from '../../../shared/src/reducers/fadersReducer'
 import {
     SOCKET_SET_FX,
     SOCKET_SET_AUX_LEVEL,
     SOCKET_SET_INPUT_GAIN,
     SOCKET_SET_INPUT_SELECTOR,
-} from '../../shared/src/constants/SOCKET_IO_DISPATCHERS'
+} from '../../../shared/src/constants/SOCKET_IO_DISPATCHERS'
 import ReductionMeter from './ReductionMeter'
 import ClassNames from 'classnames'
-import { fxParamsList } from '../../shared/src/constants/MixerProtocolInterface'
-import { IChannel } from '../../shared/src/reducers/channelsReducer'
+import { fxParamsList } from '../../../shared/src/constants/MixerProtocolInterface'
+import { IChannel } from '../../../shared/src/reducers/channelsReducer'
 import { getFaderLabel } from '../utils/labels'
 import ChanStripEq from './ChanStripEq'
