+++ client/src/components/Channel.tsx
@@ -8,14 +8,14 @@ import '../assets/css/NoUiSlider.css'
 
 //assets:
 import '../assets/css/Channel.css'
-import * as IO from '../../shared/src/constants/SOCKET_IO_DISPATCHERS'
-import { IChannelReference, IFader } from '../../shared/src/reducers/fadersReducer'
-import { ISettings } from '../../shared/src/reducers/settingsReducer'
-import { storeShowChanStrip } from '../../shared/src/actions/settingsActions'
+import * as IO from '../../../shared/src/constants/SOCKET_IO_DISPATCHERS'
+import { IChannelReference, IFader } from '../../../shared/src/reducers/fadersReducer'
+import { ISettings } from '../../../shared/src/reducers/settingsReducer'
+import { storeShowChanStrip } from '../../../shared/src/actions/settingsActions'
 import { withTranslation } from 'react-i18next'
-import { VuLabelConversionType } from '../../shared/src/constants/MixerProtocolInterface'
-import { Conversions } from '../../shared/src/actions/utils/dbConversion'
+import { VuLabelConversionType } from '../../shared../../../shared/src/constants/MixerProtocolInterface'
 import { getFaderLabel } from '../utils/labels'
+import { Conversions } from '../../../shared/src/actions/utils/dbConversion'
 
 interface IChannelInjectProps {
     t: any
