+++ client/src/components/ChannelRouteSettings.tsx
@@ -4,10 +4,10 @@ import ClassNames from 'classnames'
 import '../assets/css/ChannelRouteSettings.css'
 import { Store } from 'redux'
 import { connect } from 'react-redux'
-import { storeShowOptions } from '../../shared/src/actions/settingsActions'
-import { SOCKET_SET_ASSIGNED_FADER } from '../../shared/src/constants/SOCKET_IO_DISPATCHERS'
-import { IchMixerConnection } from '../../shared/src/reducers/channelsReducer'
-import { IFader } from '../../shared/src/reducers/fadersReducer'
+import { storeShowOptions } from '../../../shared/src/actions/settingsActions'
+import { SOCKET_SET_ASSIGNED_FADER } from '../../../shared/src/constants/SOCKET_IO_DISPATCHERS'
+import { IchMixerConnection } from '../../../shared/src/reducers/channelsReducer'
+import { IFader } from '../../../shared/src/reducers/fadersReducer'
 import { getFaderLabel } from '../utils/labels'
 
 interface IChannelSettingsInjectProps {
