+++ client/src/components/MiniChanStrip.tsx
@@ -4,8 +4,8 @@ import ReactSlider from 'react-slider'
 import '../assets/css/MiniChanStrip.css'
 import { Store } from 'redux'
 import { connect } from 'react-redux'
-import { IFader } from '../../shared/src/reducers/fadersReducer'
-import { SOCKET_SET_AUX_LEVEL } from '../../shared/src/constants/SOCKET_IO_DISPATCHERS'
+import { IFader } from '../../../shared/src/reducers/fadersReducer'
+import { SOCKET_SET_AUX_LEVEL } from '../../../shared/src/constants/SOCKET_IO_DISPATCHERS'
 import { getFaderLabel } from '../utils/labels'
 
 interface IChanStripInjectProps {
