+++ client/src/components/MiniChannels.tsx
@@ -3,9 +3,9 @@ import { connect } from 'react-redux'
 
 import '../assets/css/MiniChannels.css'
 import { Store } from 'redux'
-import { IFader } from '../../shared/src/reducers/fadersReducer'
-import { IChannels } from '../../shared/src/reducers/channelsReducer'
-import { ISettings } from '../../shared/src/reducers/settingsReducer'
+import { IFader } from '../../../shared/src/reducers/fadersReducer'
+import { IChannels } from '../../../shared/src/reducers/channelsReducer'
+import { ISettings } from '../../../shared/src/reducers/settingsReducer'
 import MiniChannel from './MiniChannel'
 import MiniChanStrip from './MiniChanStrip'
