+++ client/src/components/ReductionMeter.tsx
