+++ client/src/components/VuMeter.tsx
